pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-advisor-3000"
version = "0.1.0"
description = "Trading Advisor 3000 shell and product-plane scaffolding"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6,<7",
  "vectorbt>=0.28.5,<0.29",
  "pandas-ta-classic>=0.4.47,<0.5",
]

[project.optional-dependencies]
runtime-api = [
  "fastapi>=0.110,<1",
  "uvicorn>=0.30,<1",
  "psycopg[binary]>=3.2,<4",
]
data-proof = [
  "pyarrow>=17,<22",
  "deltalake>=0.21,<1",
  "dagster>=1.8,<2",
  "dagster-webserver>=1.8,<2",
  "pandas>=2.2,<3",
  "pandas-ta-classic>=0.3,<1",
  "vectorbt>=0.26,<1",
  "pyspark>=3.5,<4",
  "delta-spark>=3.3,<4",
  "jdk4py>=17,<22",
]
proof-docker = [
  "docker>=7,<8",
]
dev-test = [
  "pytest>=9,<10",
  "httpx>=0.27,<1",
]
dev = [
  "fastapi>=0.110,<1",
  "uvicorn>=0.30,<1",
  "psycopg[binary]>=3.2,<4",
  "pyarrow>=17,<22",
  "deltalake>=0.21,<1",
  "dagster>=1.8,<2",
  "dagster-webserver>=1.8,<2",
  "pandas>=2.2,<3",
  "pandas-ta-classic>=0.3,<1",
  "vectorbt>=0.26,<1",
  "pyspark>=3.5,<4",
  "delta-spark>=3.3,<4",
  "jdk4py>=17,<22",
  "docker>=7,<8",
  "pytest>=9,<10",
  "httpx>=0.27,<1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "--basetemp=.tmp/pytest"
